feat: change ports option to a number
BREAKING CHANGE:
Ports option was not really useful.
Allowing multiple ports to be open in a container doesn't have much use.
The idea of micro services is to have one job and one job only.

Changes:
- "ports" is renamed to "port"
- "port" can be used to specify a value on the container side.
Use this for containers that do not follow the `env.PORT` convention for their main entrypoint.